requirements.txt
colorama
pyshark
pyyaml
tqdm
xlsxwriter